refactor: 게시글 필터 조회 쿼리 수정
